Rename tensormodifyupdate.{h,cpp} -> tensor_modify_update.{h,cpp}.
