+++ document/src/vespa/document/update/tensor_modify_update.cpp
@@ -1,6 +1,6 @@
 // Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-#include "tensormodifyupdate.h"
+#include "tensor_modify_update.h"
 #include <vespa/document/base/exceptions.h>
 #include <vespa/document/base/field.h>
 #include <vespa/document/fieldvalue/document.h>
